WIP: Add additional checkbox for #1.
Pausing the execution with `Draw` button when checkbox is checked works.
Since, execution can be paused anywhere, the stackframe changes and
variables are lost.
